API Overhaul
I rearranged most of the functions into AllGroups and AllUsers, and
adjusted a lot of the API's. I think this is a more sane way of doing
things, but comments are always appreciated.

I removed the Iterators and replaced them with the "get" methods.

Also lots of docs work. I think it should be fairly straightforward to
use this crate now.

@jackpot51 I also fixed the bug where empty passwords will accept
anything.