Polish API
A full list of accomplishments for this commit can be viewed on my
message on the PR to redox_users. My main difficulty is figuring
out how to save state for AllUsers and AllGroups (Question of wether
to simply write on Drop or add a call for it). 5 other TODO's can be
found.

@goyox86 I'd like to know what you think about all this.