Timout, Iters, Bugfix
Here I did two major things:

AllGroups and AllUsers now have `iter()` and `iter_mut()` to iterate
over them. Not really sure if this is the correct way to implement
iteration here, so advice would be really nice if it's not.

Implemented a timeout when passwords are incorrect. All we do is sleep
if the password is not verified.

Also fixed a little bug with Group::to_string()