Implement Shadowfile; Fix Tests
This is still very much a WIP. There are a number of things that I know
shouldn't work, and that I haven't written tests for yet. That'll be
coming at a later date. However, the current test suite passes and it
does mostly work.

The internal API is kinda messy. I mean,
`Option<(String, Option<Encoded>)>??!! Anyway, there's some
stuff that is really not pretty, and I would be happy for advice on how
to make it look a little nicer and be more maintainable..