Polish Impl; Docs; Tests
This is at a point where the implementation should work. The only way to
know for sure that it doesn't is to write more tests :/

I basically went through all the documentation and checked it to make
sure it was up to date with the implementations, and I added some crate
level notes.

Wrote a few more tests, made sure that doing dumb things would fail,
etc.