Make argon2 optional
Resolves #26.

Since dirs is a reverse dependency of this crate and is used an awful
lot, it's nice to have as small a dependency footprint as possible. dirs
doesn't need the authentication stuff so we can drop a large number of
deps. I imagine this will probably be useful in other contexts as well.