Remove dynamic auth checking
I never liked how I had to pass around a bool and ensure that the user
had used the API correctly and panic if they hadn't. This commit adds a
marker type to `User` and `AllUsers` that ensures that a user of the
library cannot use authentication and other password related methods
unless they've instantiated the user with authentication information.

Got rid of the tests that ensured panics for incorrect usage of the API;
incorrect API use is a compile-time error now.

I removed `FromStr` for `User` and added it as a private method, later
commits will do the same for `Group` and possibly transition away from
`Display`.

Documentation in the works.