GroupBuilder and string checks
Added a GroupBuilder builder pattern for constructing groups instead of
passing the elements as a list of things to AllGroups::add_group. Fields
are verified as allowed names when added to the AllGroups.

All strings are now checked for semicolons before being written to the
database files in `/etc`. Error messages to match. This should prevent
any misuse of the `User`/`Group` fields from yielding a bad system
state.